fix checkshard bug

+++ server.sh
@@ -1,3 +1,7 @@
-sudo rm -rf .dexm*
-go build
-sudo ./dexmd sn satoshi3
+if [ "$1" != "" ]; then
+    sudo rm -rf .dexm*
+    go build
+    sudo ./dexmd sn $1
+else
+    echo "Wallet empty"
+fi
